Issue #681 Move GroundWaterFlowModel and GroundWaterTransportModels to separate files
This commit extracts the GroundWaterFlowModel and the GroundWaterTransportModels from the model.py file and moves them to seprate classes
