+++ imod/mf6/__init__.py
@@ -31,7 +31,8 @@ from imod.mf6.ims import (
 )
 from imod.mf6.ist import ImmobileStorageTransfer
 from imod.mf6.lak import Lake, LakeData, OutletManning, OutletSpecified, OutletWeir
-from imod.mf6.model import GroundwaterFlowModel, GroundwaterTransportModel
+from imod.mf6.model_gwf import GroundwaterFlowModel
+from imod.mf6.model_gwt import GroundwaterTransportModel
 from imod.mf6.mst import MobileStorageTransfer
 from imod.mf6.npf import NodePropertyFlow
 from imod.mf6.oc import OutputControl
