+++ imod/mf6/model.py
@@ -6,7 +6,7 @@ import inspect
 import pathlib
 from copy import deepcopy
 from pathlib import Path
-from typing import Dict, List, Optional, Tuple, Union
+from typing import Dict, Optional, Tuple, Union
 
 import cftime
 import jinja2
@@ -18,7 +18,6 @@ import xugrid as xu
 from jinja2 import Template
 
 import imod
-from imod.mf6.clipped_boundary_condition_creator import create_clipped_boundary
 from imod.mf6.package import Package
 from imod.mf6.regridding_utils import RegridderInstancesCollection, RegridderType
 from imod.mf6.statusinfo import NestedStatusInfo, StatusInfo, StatusInfoBase
@@ -356,7 +355,35 @@ class Modflow6Model(collections.UserDict, abc.ABC):
         y_max: Optional[float] = None,
         state_for_boundary: Optional[GridDataArray] = None,
     ):
-        raise NotImplementedError
+        """
+        Clip a model by a bounding box (time, layer, y, x).
+
+        Slicing intervals may be half-bounded, by providing None:
+
+        * To select 500.0 <= x <= 1000.0:
+          ``clip_box(x_min=500.0, x_max=1000.0)``.
+        * To select x <= 1000.0: ``clip_box(x_min=None, x_max=1000.0)``
+          or ``clip_box(x_max=1000.0)``.
+        * To select x >= 500.0: ``clip_box(x_min = 500.0, x_max=None.0)``
+          or ``clip_box(x_min=1000.0)``.
+
+        Parameters
+        ----------
+        time_min: optional
+        time_max: optional
+        layer_min: optional, int
+        layer_max: optional, int
+        x_min: optional, float
+        x_max: optional, float
+        y_min: optional, float
+        y_max: optional, float
+        state_for_boundary :
+        """
+        clipped = self._clip_box_packages(
+            time_min, time_max, layer_min, layer_max, x_min, x_max, y_min, y_max
+        )
+
+        return clipped
 
     def _clip_box_packages(
         self,
@@ -525,168 +552,3 @@ class Modflow6Model(collections.UserDict, abc.ABC):
 
     def is_use_newton(self):
         return False
-
-
-class GroundwaterFlowModel(Modflow6Model):
-    _mandatory_packages = ("npf", "ic", "oc", "sto")
-    _model_id = "gwf6"
-
-    def __init__(
-        self,
-        listing_file: str = None,
-        print_input: bool = False,
-        print_flows: bool = False,
-        save_flows: bool = False,
-        newton: bool = False,
-        under_relaxation: bool = False,
-    ):
-        super().__init__()
-        self._options = {
-            "listing_file": listing_file,
-            "print_input": print_input,
-            "print_flows": print_flows,
-            "save_flows": save_flows,
-            "newton": newton,
-            "under_relaxation": under_relaxation,
-        }
-        self._template = initialize_template("gwf-nam.j2")
-
-    def _get_unique_regridder_types(self) -> Dict[RegridderType, str]:
-        """
-        This function loops over the packages and  collects all regridder-types that are in use.
-        Differences in associated functions are ignored. It focusses only on the types. So if a
-        model uses both Overlap(mean) and Overlap(harmonic_mean), this function will return just one
-        Overlap regridder:  the first one found, in this case Overlap(mean)
-        """
-        methods = {}
-        for pkg_name, pkg in self.items():
-            if pkg.is_regridding_supported():
-                pkg_methods = pkg.get_regrid_methods()
-                for variable in pkg_methods:
-                    if (
-                        variable in pkg.dataset.data_vars
-                        and pkg.dataset[variable].values[()] is not None
-                    ):
-                        regriddertype = pkg_methods[variable][0]
-                        if regriddertype not in methods.keys():
-                            functiontype = pkg_methods[variable][1]
-                            methods[regriddertype] = functiontype
-            else:
-                raise NotImplementedError(
-                    f"regridding is not implemented for package {pkg_name} of type {type(pkg)}"
-                )
-        return methods
-
-    def clip_box(
-        self,
-        time_min: Optional[str] = None,
-        time_max: Optional[str] = None,
-        layer_min: Optional[int] = None,
-        layer_max: Optional[int] = None,
-        x_min: Optional[float] = None,
-        x_max: Optional[float] = None,
-        y_min: Optional[float] = None,
-        y_max: Optional[float] = None,
-        state_for_boundary: Optional[GridDataArray] = None,
-    ):
-        clipped = super()._clip_box_packages(
-            time_min, time_max, layer_min, layer_max, x_min, x_max, y_min, y_max
-        )
-
-        clipped_boundary_condition = self.__create_boundary_condition_clipped_boundary(
-            self, clipped, state_for_boundary
-        )
-        if clipped_boundary_condition is not None:
-            clipped["chd_clipped"] = clipped_boundary_condition
-
-        return clipped
-
-    def __create_boundary_condition_clipped_boundary(
-        self,
-        original_model: Modflow6Model,
-        clipped_model: Modflow6Model,
-        state_for_boundary: Optional[GridDataArray],
-    ):
-        unassigned_boundary_original_domain = (
-            self.__create_boundary_condition_for_unassigned_boundary(
-                original_model, state_for_boundary
-            )
-        )
-
-        return self.__create_boundary_condition_for_unassigned_boundary(
-            clipped_model, state_for_boundary, [unassigned_boundary_original_domain]
-        )
-
-    @staticmethod
-    def __create_boundary_condition_for_unassigned_boundary(
-        model: Modflow6Model,
-        state_for_boundary: Optional[GridDataArray],
-        additional_boundaries: Optional[List[imod.mf6.ConstantHead]] = None,
-    ):
-        if state_for_boundary is None:
-            return None
-
-        constant_head_packages = [
-            pkg for name, pkg in model.items() if isinstance(pkg, imod.mf6.ConstantHead)
-        ]
-
-        additional_boundaries = [
-            item for item in additional_boundaries or [] if item is not None
-        ]
-
-        constant_head_packages.extend(additional_boundaries)
-
-        return create_clipped_boundary(
-            model.domain, state_for_boundary, constant_head_packages
-        )
-
-    def is_use_newton(self):
-        return self._options["newton"]
-
-    def set_newton(self, is_newton: bool) -> None:
-        self._options["newton"] = is_newton
-
-
-class GroundwaterTransportModel(Modflow6Model):
-    """
-    The GroundwaterTransportModel (GWT) simulates transport of a single solute
-    species flowing in groundwater.
-    """
-
-    _mandatory_packages = ("mst", "dsp", "oc", "ic")
-    _model_id = "gwt6"
-
-    def __init__(
-        self,
-        listing_file: str = None,
-        print_input: bool = False,
-        print_flows: bool = False,
-        save_flows: bool = False,
-    ):
-        super().__init__()
-        self._options = {
-            "listing_file": listing_file,
-            "print_input": print_input,
-            "print_flows": print_flows,
-            "save_flows": save_flows,
-        }
-
-        self._template = initialize_template("gwt-nam.j2")
-
-    def clip_box(
-        self,
-        time_min: str = None,
-        time_max: str = None,
-        layer_min: int = None,
-        layer_max: int = None,
-        x_min: float = None,
-        x_max: float = None,
-        y_min: float = None,
-        y_max: float = None,
-        state_for_boundary: GridDataArray = None,
-    ):
-        clipped = super()._clip_box_packages(
-            time_min, time_max, layer_min, layer_max, x_min, x_max, y_min, y_max
-        )
-
-        return clipped
