+++ imod/mf6/model_gwt.py
@@ -1,0 +1,30 @@
+from __future__ import annotations
+
+from imod.mf6.model import Modflow6Model, initialize_template
+
+
+class GroundwaterTransportModel(Modflow6Model):
+    """
+    The GroundwaterTransportModel (GWT) simulates transport of a single solute
+    species flowing in groundwater.
+    """
+
+    _mandatory_packages = ("mst", "dsp", "oc", "ic")
+    _model_id = "gwt6"
+
+    def __init__(
+        self,
+        listing_file: str = None,
+        print_input: bool = False,
+        print_flows: bool = False,
+        save_flows: bool = False,
+    ):
+        super().__init__()
+        self._options = {
+            "listing_file": listing_file,
+            "print_input": print_input,
+            "print_flows": print_flows,
+            "save_flows": save_flows,
+        }
+
+        self._template = initialize_template("gwt-nam.j2")
